+++ interfaces/vsprng.h
@@ -35,7 +35,7 @@ class VSPRNG
     virtual SIMD_FLT get_rn_flt() = 0;
     virtual SIMD_DBL get_rn_dbl() = 0;
     virtual SIMD_INT get_seed_rng() const = 0;
-    virtual unsigned long int get_ngens() const = 0;
+    virtual int get_ngens() const = 0;
 #if defined(DEBUG)
     virtual SIMD_INT get_seed() const = 0;
     virtual SIMD_INT get_prime() const = 0;
